refactor: divide the application into components and add initial functionality
Divide the application into 5 main components (App, Book, BookShelf, ListBooks, and SearchBooks)

App Component:
Stateful component holding the books state, and updating it. Routing between SearchBooks, and ListBooks components.

Book Component:
Functional stateless component responsible for rendering the book item. It should takes a book object as an input, and fire an event in case of book update.

BookShelf Component:
Functional stateless component responsible for rendering the books in the shelf. It should takes a books array as an input, and fire an event in case of book update.

ListBooks Component:
Functional stateless component responsible for rendering the three selves. It should takes a books array as an input, and fire an event in case of book update.

SearchBooks Component:
Stateful component holding the books that the user is searching for. It should takes a books array as an input, and fire an event in case of book update.